Support building boot2docker.iso for Parallels Desktop
This adds Parallels Tools support in the boot2docker.iso. With it it is
possible to use the prl_fs kernel module to mount the hosts
filesystem.

This requires that you have Parallels Desktop for Mac installed and the
built boot2docker.iso WILL CONTAIN PROPRIETARY SOFTWARE WHICH IS NOT
FREE TO BE DISTRIBUTED under the current EULA!

See doc/BUILD.Parallels.md